Refactor HTML and migrate Alpine.js components 🚀🎉
- Removed `alpine-integration.js` and split its code into multiple dedicated components for better organization: `alpine-auth.js`, `alpine-theme.js`, `alpine-posts.js`, and `alpine-top8.js`.
- Introduced a new `index.js` file as the main entry point for Alpine, replacing multiple script tags in `index.html` with a single modular import. This significantly tidies up the HTML file!
- Added a new folder for Alpine files and placed them accordingly to enhance maintainability 🌱.
- Ensured that old script tags in `index.html` related to Alpine were replaced with appropriate imports, avoiding the need for a massive script file 🤯.
- Modified `index.html` to establish cleaner links to CSS and avoid redundancy in including the same stylesheet twice.
- Created a basic `package.json` for better dependency management and build processes using Parcel. This includes scripts for clean, format, lint, and build commands for your coding delight 🛠️.

These changes serve to:
- Enhance clarity and organization of the JavaScript components.
- Allow for easier collaboration in the future as modules can be maintained or updated independently.
- Improve the overall structure, making it simpler for other developers to jump into the code 🎈!

Thank God for modular code, right?! 🎉
