+++ app/index.html
@@ -5,22 +5,11 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>Top 8 Friends</title>
 
-        <!-- External Dependencies -->
-        <!-- Alpine.js must be first and defer -->
-        <script src="//unpkg.com/alpinejs" defer></script>
+        <!-- Styles -->
+        <link rel="stylesheet" href="css/styles.css" />
 
-        <!-- Supporting classes -->
-        <script src="js/config.js"></script>
-        <script src="js/auth.js"></script>
-        <script src="js/profile.js"></script>
-        <script src="js/storage.js"></script>
-        <script src="js/posts.js"></script>
-        <script src="js/pds.js"></script>
-        <script src="js/themePDS.js"></script>
-        <script src="js/top8.js"></script>
-
-        <!-- Alpine integration must be last -->
-        <script src="js/alpine-integration.js"></script>
+        <!-- Main JS Entry Point -->
+        <script type="module" src="js/index.js"></script>
 
         <!-- Styles -->
         <link rel="stylesheet" href="css/styles.css" />
